Rather than attempt to solve problems for everyone, I made some substantial changes to solve problems for me.  For starters, I wanted this to run as a service within Systemd.  I have no interest in attempting to make sense of Heroku or the Docker implementation.
Secondly, I am a firm believer that we should have been off python 2.7 ages ago.  I made changes to modernize this to be Python 3.6 plus.

I updated the requirements.txt by providing our base requirements file and freezing based off that.  I can do this because I also wrote instructions
for setting up and using virtualenv.

This also stands to be a pretty useful sample for me when it comes to writing slash commands in the future.

There is still work to be done, in specific the fact that we're using the flask development server and should be using uwsgi instead.  I will
make it a point to get that in place in the near days to come.